Use internal markers also for generated meshes. Refactor code to use _utils.mesh
